crawler/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
python-dateutil>=2.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
